pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-keyword-analyzer"
version = "7.0.0"
description = "Geo keyword analyzer with async crawl, dedupe, dashboard and reports."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12",
  "certifi>=2024.0.0",
  "curl_cffi>=0.6.0",
  "httpx>=0.27.0",
  "playwright>=1.50.0",
  "playwright-stealth>=1.0.6",
  "matplotlib>=3.8.0",
  "networkx>=3.2",
  "pyahocorasick>=2.0.0",
  "PyYAML>=6.0.1",
  "requests>=2.31.0",
  "tqdm>=4.66.0",
  "trafilatura>=1.9.0",
  "urllib3>=2.2.0",
  "wordcloud>=1.9.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0"]

[tool.setuptools]
packages = ["geo_analyzer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
